requirements.txt
tensorflow-macos
tensorflow-metal
opencv-python
numpy
matplotlib
scikit-learn
pillow
tqdm
mediapipe
